refactor: slightly improve network related files
